pyproject.toml
[project]
name = "excel-sheet-matching-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-documentintelligence>=1.0.2",
    "langchain>=0.3.24",
    "langchain-community>=0.3.23",
    "langchain-core>=0.3.56",
    "langchain-google-genai>=2.1.3",
    "langchain-openai>=0.3.14",
    "langfuse>=2.60.3",
    "langgraph>=0.4.1",
    "opencc>=1.1.9",
    "openpyxl>=3.1.5",
    "pydantic>=2.11.4",
    "pymupdf>=1.25.5",
    "python-dotenv>=1.1.0",
]


